Add Dar Es Salaam
I know this is an unlikely candidate destination for the list, but having traveled to many places and eaten numerous pizzas - this is a standout. Get the basic cheese and add the local variant of the Scotchbonnet pepper added on top (pilipili mbuzi). Enjoy with a beverage of your chocie.